ADDFILE should never make first block sparse
When checking if a file block can be sparse (i.e. not actually
allocated on disk), ensure the test always fails for the first
block of the file, as this would violate the ProDOS file system
specification, and makes ProDOS-8 unhappy.

Additionally, when loading a volume, log if files are encountered
that have sparse first blocks.

https://github.com/mach-kernel/cadius/issues/42